+++ queue/queue_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 	"testing"
+	"time"
 )
 
 func TestQueuePush(t *testing.T) {
@@ -29,7 +30,7 @@ func TestQueuePush(t *testing.T) {
 
 	sm.EXPECT().PushJobToQueue(queueName, que, j).Return(nil)
 	sm.EXPECT().CollectInFlightJobNumberQueue(queueName).AnyTimes()
-	q := New(sm, mLog, queueName)
+	q := New(sm, mLog, queueName, time.Duration(0))
 
 	err = q.Push(j)
 	require.NoError(t, err)
@@ -45,12 +46,43 @@ func TestQueuePop(t *testing.T) {
 	expectNV := job.NameVersion("haha")
 	sm := storemock.NewMockStore(ctrl)
 	mLog := logmock.NewMockLogger(ctrl)
+	blockTime := time.Duration(0)
 
-	sm.EXPECT().PopJobFromQueue(que).Return(expectNV, nil)
+	sm.EXPECT().PopJobFromQueue(que, blockTime).Return(expectNV, nil)
 	sm.EXPECT().CollectInFlightJobNumberQueue(queueName).AnyTimes()
-	q := New(sm, mLog, queueName)
+	q := New(sm, mLog, queueName, blockTime)
 
 	nv, err := q.Pop(jobTopic)
 	require.NoError(t, err)
 	require.Equal(t, expectNV, nv)
+}
+
+func TestQueuePopWithBlockTime(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+	queueName := "test_queue_name"
+	jobTopic := job.Topic("job_topic")
+	que := fmt.Sprintf("%s_%s", queueName, jobTopic)
+
+	expectNV := job.NameVersion("haha")
+	sm := storemock.NewMockStore(ctrl)
+	mLog := logmock.NewMockLogger(ctrl)
+	blockTime := 2 * time.Second
+
+	sm.EXPECT().PopJobFromQueue(que, blockTime).DoAndReturn(
+		func(queue string, blockTime time.Duration) (job.NameVersion, error) {
+			time.Sleep(blockTime)
+			return expectNV, nil
+		},
+	)
+	sm.EXPECT().CollectInFlightJobNumberQueue(queueName).AnyTimes()
+	q := New(sm, mLog, queueName, blockTime)
+
+	startTime := time.Now()
+	nv, err := q.Pop(jobTopic)
+	dur := time.Since(startTime)
+
+	require.NoError(t, err)
+	require.Equal(t, expectNV, nv)
+	require.GreaterOrEqual(t, dur, blockTime)
 }