+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
 
 1. Add more unit testings.
-2. Add prometheus metrics api
+2. Add prometheus metrics api.
+3. Pop Api is using BRPOP to pop job from redis queue.